fix: provider configuration validation
Updates the provider configuration to address the issue where the function assumes the configuration is for Cloud Builder if the SDDC Manager username is not provided.

Added a check to ensure that at least one of the configurations is provided. If neither is provided, we can return an appropriate error message.

Ref #238

Signed-off-by: Ryan Johnson <ryan.johnson@broadcom.com>